feat: agora pode editar a categoria e excluir a categoria

+++ back-end/src/category/category.service.ts
@@ -9,7 +9,7 @@ export class CategoryService {
 	constructor(private readonly prisma: PrismaService) {}
 
   async create(createCategoryDto: CreateCategoryDto, authenticatedUser: JwtUserPayload) {
-    if(createCategoryDto.userId !== authenticatedUser.id && authenticatedUser.role !== 'ADMIN'){
+    if(createCategoryDto.userId !== authenticatedUser.id){
       throw new ForbiddenException('Acesso negado. Não é permitido utilizar o ID de outro usuário.')
     }
 
@@ -29,84 +29,67 @@ export class CategoryService {
 	}
 
 	async findAll(authenticatedUser: JwtUserPayload) {
-		// ADMINs conseguem ver as categorias de TODOS os usuários
-		// USERs conseguem ver somente suas categorias
-
-		if (authenticatedUser.role == 'ADMIN') {
-			return this.prisma.category.findMany();
-		}
-
-		// limita as categorias ao usuário
 		return this.prisma.category.findMany({
 			where: { userId: authenticatedUser.id },
 		});
 	}
 
 	async findOne(id: string, authenticatedUser: JwtUserPayload) {
-		// ADMINs podem ver qualquer categoria
-		// USERs podem ver somente as suas
-
-		// limita as categorias ao usuário
-		var existingCategory = await this.prisma.category.findUnique({
+		const existingCategory = await this.prisma.category.findUnique({
 			where: { id, userId: authenticatedUser.id },
 		});
 
-		if (authenticatedUser.role == 'ADMIN') {
-			var existingCategory = await this.prisma.category.findUnique({
-				where: { id },
-			});
-		}
-
 		if (!existingCategory) throw new NotFoundException('Categoria não encontrada.');
-		else if (existingCategory.userId !== authenticatedUser.id && authenticatedUser.role !== 'ADMIN')
-			throw new ForbiddenException('Acesso negado. Permissão insuficiente.');
 
 		return existingCategory;
 	}
 
 	async update(id: string, updateCategoryDto: UpdateCategoryDto, authenticatedUser: JwtUserPayload) {
-		// ADMINs dão update em qualquer categoria
-		// USERs dão update só em suas categorias
-
-		// limita as categorias ao usuário
-		var existingCategory = await this.prisma.category.findUnique({
+		const existingCategory = await this.prisma.category.findUnique({
 			where: { id, userId: authenticatedUser.id },
 		});
 
-		if (authenticatedUser.role == 'ADMIN') {
-			var existingCategory = await this.prisma.category.findUnique({
-				where: { id },
-			});
-		}
-
 		if (!existingCategory) throw new NotFoundException('Categoria não encontrada.');
-		else if (authenticatedUser.id !== existingCategory.userId && authenticatedUser.role !== 'ADMIN')
-			throw new ForbiddenException('Acesso negado. Permissão insuficiente.');
+
+		const categoryWithSameName = await this.prisma.category.findFirst({
+			where: {
+				userId: authenticatedUser.id,
+				name: updateCategoryDto.name,
+				id: { not: id },
+			},
+		});
+
+		if (categoryWithSameName) {
+			throw new ConflictException('Categoria já existente. Caso queira criar uma nova, utilize outro nome.');
+		}
 
 		return this.prisma.category.update({
 			where: { id },
-			data: updateCategoryDto,
+			data: {
+				name: updateCategoryDto.name,
+				color: updateCategoryDto.color,
+				categoryType: updateCategoryDto.categoryType,
+			},
 		});
 	}
 
 	async remove(id: string, authenticatedUser: JwtUserPayload) {
-		// ADMINs apagam tudo
-		// USERs apagam só o seu
-
-		// limita as categorias ao usuário
-		var existingCategory = await this.prisma.category.findUnique({
+		const existingCategory = await this.prisma.category.findUnique({
 			where: { id, userId: authenticatedUser.id },
 		});
 
-		if (authenticatedUser.role == 'ADMIN') {
-			var existingCategory = await this.prisma.category.findUnique({
-				where: { id },
-			});
-		}
-
 		if (!existingCategory) throw new NotFoundException('Categoria não encontrada.');
-		else if (authenticatedUser.id !== existingCategory.userId && authenticatedUser.role !== 'ADMIN')
-			throw new ForbiddenException('Acesso negado. Permissão insuficiente.');
+
+		const linkedLaunchesCount = await this.prisma.launch.count({
+			where: {
+				categoryId: id,
+				userId: authenticatedUser.id,
+			},
+		});
+
+		if (linkedLaunchesCount > 0) {
+			throw new ConflictException('Não é possível excluir uma categoria que possui lançamentos vinculados.');
+		}
 
 		return this.prisma.category.delete({
 			where: { id },
